refactor(endpoint): move MergeEndpoints from source to endpoint package
