+++ endpoint/utils_test.go
@@ -19,9 +19,12 @@ package endpoint
 import (
 	"testing"
 
+	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	logtest "sigs.k8s.io/external-dns/internal/testutils/log"
 	"sigs.k8s.io/external-dns/pkg/events"
 )
 
@@ -236,4 +239,310 @@ func TestAttachRefObject(t *testing.T) {
 	for _, ep := range eps {
 		assert.Same(t, ref, ep.RefObject())
 	}
+}
+
+func TestMergeEndpoints(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    []*Endpoint
+		expected []*Endpoint
+	}{
+		{
+			name:     "nil input returns nil",
+			input:    nil,
+			expected: nil,
+		},
+		{
+			name:     "empty input returns empty",
+			input:    []*Endpoint{},
+			expected: []*Endpoint{},
+		},
+		{
+			name: "single endpoint unchanged",
+			input: []*Endpoint{
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"1.2.3.4"}},
+			},
+			expected: []*Endpoint{
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"1.2.3.4"}},
+			},
+		},
+		{
+			name: "different keys not merged",
+			input: []*Endpoint{
+				{DNSName: "a.example.com", RecordType: RecordTypeA, Targets: Targets{"1.2.3.4"}},
+				{DNSName: "b.example.com", RecordType: RecordTypeA, Targets: Targets{"5.6.7.8"}},
+			},
+			expected: []*Endpoint{
+				{DNSName: "a.example.com", RecordType: RecordTypeA, Targets: Targets{"1.2.3.4"}},
+				{DNSName: "b.example.com", RecordType: RecordTypeA, Targets: Targets{"5.6.7.8"}},
+			},
+		},
+		{
+			name: "same DNSName different RecordType not merged",
+			input: []*Endpoint{
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"1.2.3.4"}},
+				{DNSName: "example.com", RecordType: RecordTypeAAAA, Targets: Targets{"2001:db8::1"}},
+			},
+			expected: []*Endpoint{
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"1.2.3.4"}},
+				{DNSName: "example.com", RecordType: RecordTypeAAAA, Targets: Targets{"2001:db8::1"}},
+			},
+		},
+		{
+			name: "same key merged with sorted targets",
+			input: []*Endpoint{
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"5.6.7.8"}},
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"1.2.3.4"}},
+			},
+			expected: []*Endpoint{
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"1.2.3.4", "5.6.7.8"}},
+			},
+		},
+		{
+			name: "multiple endpoints same key merged",
+			input: []*Endpoint{
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"3.3.3.3"}},
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"1.1.1.1"}},
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"2.2.2.2"}},
+			},
+			expected: []*Endpoint{
+				{DNSName: "example.com", RecordType: RecordTypeA, Targets: Targets{"1.1.1.1", "2.2.2.2", "3.3.3.3"}},
+			},
+		},
+		{
+			name: "mixed merge and no merge",
+			input: []*Endpoint{
+				{DNSName: "a.example.com", RecordType: RecordTypeA, Targets: Targets{"1.1.1.1"}},
+				{DNSName: "b.example.com", RecordType: RecordTypeA, Targets: Targets{"2.2.2.2"}},
+				{DNSName: "a.example.com", RecordType: RecordTypeA, Targets: Targets{"3.3.3.3"}},
+			},
+			expected: []*Endpoint{
+				{DNSName: "a.example.com", RecordType: RecordTypeA, Targets: Targets{"1.1.1.1", "3.3.3.3"}},
+				{DNSName: "b.example.com", RecordType: RecordTypeA, Targets: Targets{"2.2.2.2"}},
+			},
+		},
+		{
+			name: "duplicate targets deduplicated",
+			input: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4", "1.2.3.4", "5.6.7.8"),
+			},
+			expected: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4", "5.6.7.8"),
+			},
+		},
+		{
+			name: "duplicate targets across merged endpoints deduplicated",
+			input: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4"),
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4", "5.6.7.8"),
+			},
+			expected: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4", "5.6.7.8"),
+			},
+		},
+		{
+			name: "CNAME endpoints not merged",
+			input: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com"),
+				NewEndpoint("example.com", RecordTypeCNAME, "b.elb.com"),
+			},
+			expected: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com"),
+				NewEndpoint("example.com", RecordTypeCNAME, "b.elb.com"),
+			},
+		},
+		{
+			name: "CNAME with no targets is skipped",
+			input: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeCNAME),
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4"),
+			},
+			expected: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4"),
+			},
+		},
+		{
+			name: "identical CNAME endpoints deduplicated",
+			input: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com"),
+				NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com"),
+			},
+			expected: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com"),
+			},
+		},
+		{
+			name: "same key with different TTL not merged",
+			input: []*Endpoint{
+				NewEndpointWithTTL("example.com", RecordTypeA, 300, "1.2.3.4"),
+				NewEndpointWithTTL("example.com", RecordTypeA, 600, "5.6.7.8"),
+			},
+			expected: []*Endpoint{
+				NewEndpointWithTTL("example.com", RecordTypeA, 300, "1.2.3.4"),
+				NewEndpointWithTTL("example.com", RecordTypeA, 600, "5.6.7.8"),
+			},
+		},
+		{
+			name: "same DNSName and RecordType with different SetIdentifier not merged",
+			input: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4").WithSetIdentifier("us-east-1"),
+				NewEndpoint("example.com", RecordTypeA, "5.6.7.8").WithSetIdentifier("eu-west-1"),
+			},
+			expected: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4").WithSetIdentifier("us-east-1"),
+				NewEndpoint("example.com", RecordTypeA, "5.6.7.8").WithSetIdentifier("eu-west-1"),
+			},
+		},
+		{
+			name: "same DNSName, RecordType and SetIdentifier targets are merged",
+			input: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4").WithSetIdentifier("us-east-1"),
+				NewEndpoint("example.com", RecordTypeA, "5.6.7.8").WithSetIdentifier("us-east-1"),
+			},
+			expected: []*Endpoint{
+				NewEndpoint("example.com", RecordTypeA, "1.2.3.4", "5.6.7.8").WithSetIdentifier("us-east-1"),
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := MergeEndpoints(tt.input)
+			assert.ElementsMatch(t, tt.expected, result)
+		})
+	}
+}
+
+func TestMergeEndpoints_RefObjects(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    func() []*Endpoint
+		expected func(*testing.T, []*Endpoint)
+	}{
+		{
+			name:  "empty input",
+			input: func() []*Endpoint { return []*Endpoint{} },
+			expected: func(t *testing.T, ep []*Endpoint) {
+				assert.Empty(t, ep)
+			},
+		},
+		{
+			name: "single endpoint",
+			input: func() []*Endpoint {
+				return []*Endpoint{
+					NewEndpoint("example.com", RecordTypeA, "1.2.3.4").WithRefObject(
+						events.NewObjectReference(&v1.Service{
+							ObjectMeta: metav1.ObjectMeta{Name: "foo", Namespace: "default", UID: "123"},
+						}, "service")),
+				}
+			},
+			expected: func(t *testing.T, ep []*Endpoint) {
+				assert.Len(t, ep, 1)
+				assert.Equal(t, "service", ep[0].RefObject().Source)
+				assert.Equal(t, "foo", ep[0].RefObject().Name)
+				assert.Equal(t, "123", string(ep[0].RefObject().UID))
+			},
+		},
+		{
+			name: "two endpoints merged and only single refObject preserved",
+			input: func() []*Endpoint {
+				return []*Endpoint{
+					NewEndpoint("a.example.com", RecordTypeA, "1.1.1.1").WithRefObject(
+						events.NewObjectReference(&v1.Service{
+							ObjectMeta: metav1.ObjectMeta{Name: "foo", Namespace: "default", UID: "123"},
+						}, "service")),
+					NewEndpoint("a.example.com", RecordTypeA, "1.1.1.1").WithRefObject(
+						events.NewObjectReference(&v1.Service{
+							ObjectMeta: metav1.ObjectMeta{Name: "bar", Namespace: "ns", UID: "345"},
+						}, "service")),
+				}
+			},
+			expected: func(t *testing.T, ep []*Endpoint) {
+				assert.Len(t, ep, 1)
+				assert.Equal(t, "service", ep[0].RefObject().Source)
+				assert.Equal(t, "foo", ep[0].RefObject().Name)
+				assert.Equal(t, "123", string(ep[0].RefObject().UID))
+				assert.NotEqual(t, "345", string(ep[0].RefObject().UID))
+			},
+		},
+		{
+			name: "two endpoints not merged and two refObject preserved",
+			input: func() []*Endpoint {
+				return []*Endpoint{
+					NewEndpoint("a.example.com", RecordTypeA, "1.1.1.1").WithRefObject(
+						events.NewObjectReference(&v1.Service{
+							ObjectMeta: metav1.ObjectMeta{Name: "foo", Namespace: "default", UID: "123"},
+						}, "service")),
+					NewEndpoint("b.example.com", RecordTypeA, "1.1.1.2").WithRefObject(
+						events.NewObjectReference(&v1.Service{
+							ObjectMeta: metav1.ObjectMeta{Name: "bar", Namespace: "ns", UID: "345"},
+						}, "service")),
+				}
+			},
+			expected: func(t *testing.T, ep []*Endpoint) {
+				assert.Len(t, ep, 2)
+				assert.NotEqual(t, ep[0], ep[1])
+				for _, el := range ep {
+					assert.Equal(t, "service", el.RefObject().Source)
+					assert.Contains(t, []string{"foo", "bar"}, el.RefObject().Name)
+					assert.Contains(t, []string{"123", "345"}, string(el.RefObject().UID))
+				}
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			result := MergeEndpoints(tt.input())
+			tt.expected(t, result)
+		})
+	}
+}
+
+func TestMergeEndpointsLogging(t *testing.T) {
+	t.Run("warns on CNAME conflict", func(t *testing.T) {
+		hook := logtest.LogsUnderTestWithLogLevel(log.WarnLevel, t)
+
+		MergeEndpoints([]*Endpoint{
+			NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com"),
+			NewEndpoint("example.com", RecordTypeCNAME, "b.elb.com"),
+		})
+
+		logtest.TestHelperLogContainsWithLogLevel("Only one CNAME per name", log.WarnLevel, hook, t)
+		logtest.TestHelperLogContains("example.com CNAME a.elb.com", hook, t)
+		logtest.TestHelperLogContains("example.com CNAME b.elb.com", hook, t)
+	})
+
+	t.Run("no warning for identical CNAMEs", func(t *testing.T) {
+		hook := logtest.LogsUnderTestWithLogLevel(log.WarnLevel, t)
+
+		MergeEndpoints([]*Endpoint{
+			NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com"),
+			NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com"),
+		})
+
+		logtest.TestHelperLogNotContains("Only one CNAME per name", hook, t)
+	})
+
+	t.Run("no warning for same DNSName with different SetIdentifier", func(t *testing.T) {
+		hook := logtest.LogsUnderTestWithLogLevel(log.WarnLevel, t)
+
+		MergeEndpoints([]*Endpoint{
+			NewEndpoint("example.com", RecordTypeCNAME, "a.elb.com").WithSetIdentifier("weight-1"),
+			NewEndpoint("example.com", RecordTypeCNAME, "b.elb.com").WithSetIdentifier("weight-2"),
+		})
+
+		logtest.TestHelperLogNotContains("Only one CNAME per name", hook, t)
+	})
+
+	t.Run("debug log for CNAME with no targets", func(t *testing.T) {
+		hook := logtest.LogsUnderTestWithLogLevel(log.DebugLevel, t)
+
+		MergeEndpoints([]*Endpoint{
+			NewEndpoint("example.com", RecordTypeCNAME),
+		})
+
+		logtest.TestHelperLogContainsWithLogLevel("Skipping CNAME endpoint", log.DebugLevel, hook, t)
+		logtest.TestHelperLogContains("example.com", hook, t)
+	})
 }
